% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsky.R
\name{bsky_profile}
\alias{bsky_profile}
\title{Get a Bluesky user profile}
\usage{
bsky_profile(handle)
}
\arguments{
\item{handle}{User handle (e.g. "jay.bsky.team") or DID}
}
\value{
tibble: one row with did, handle, display_name, description,
followers_count, follows_count, posts_count, created_at
}
\description{
Get a Bluesky user profile
}
