% Generated by build script
\name{bsky_profile}
\alias{bsky_profile}
\title{Get a Bluesky user profile}
\usage{
bsky_profile()
}
\arguments{
\item{handle}{User handle (e.g. "jay.bsky.team") or DID}
}
\value{tibble: one row with did, handle, display_name, description,}
\description{Get a Bluesky user profile}
